rag-content/requirements.txt
langchain
langchain_community
pypdf
chromadb
faiss-cpu
pypdf
bs4
python-dotenv
openai
langchain-openai
nest_asyncio
gpt4all
flask
streamlit